Initial, rudemantary application interface
Instead of commiting the logic first, I decided to play around with
some GUI options to verify that this is easier for me in Rust than in
Python.

The logic will be much easier to implement, as I already have some
experience in writing a cli application in python for the same purpose.